+++ emg3d/simulations.py
@@ -1002,6 +1002,7 @@ class Simulation:
         rfield = fields.Field(grid, frequency=freq)
 
         # Loop over receivers, input as source.
+        synthetic = self.data.synthetic.loc[source, :, frequency]
         for name, rec in self.survey.receivers.items():
 
             # Get residual of this receiver.
@@ -1011,16 +1012,10 @@ class Simulation:
 
             # Residual source strength: Weighted residual, normalized by -smu0.
             weight = self.data.weights.loc[source, name, frequency].data
-            data_complex_rec = self.data.synthetic.loc[
-                source, name, frequency
-            ].data
-            data_complex_rec_deriv = rec.data_deriv(
-                data_complex_rec, adjoint=True
-            )[0]
-            strength = np.conj(
-                residual * weight *
-                data_complex_rec_deriv / -rfield.smu0
-            )
+            strength = np.conj(residual * weight / -rfield.smu0)
+
+            # Apply chain rule to strength if data_type != complex.
+            rec.derivative_chain(strength, synthetic.loc[name].data)
 
             # TODO ideally, we want here simply a single call independent of
             # receiver type, and implement all receiver-type specific things in
@@ -1116,19 +1111,10 @@ class Simulation:
             # Store responses at receivers.
             resp = self._get_responses(src, freq, out[i][0])
 
-            # Apply a chainrule for the different data_type than complex
-            # (e.g. amp)
-            data_complex_deriv = []
-            for name, rec in self.survey.receivers.items():
-                data_complex_rec = self.data.synthetic.loc[
-                    src, name, freq
-                ].data
-                data_complex_rec_deriv = rec.data_deriv(
-                    data_complex_rec,
-                    adjoint=False
-                )
-                data_complex_deriv.append(data_complex_rec_deriv)
-            resp *= np.hstack(data_complex_deriv)
+            # Apply chain rule to responses if data_type != complex.
+            synthetic = self.data.synthetic.loc[src, :, freq]
+            for ii, (name, rec) in enumerate(self.survey.receivers.items()):
+                rec.derivative_chain(resp[ii], synthetic.loc[name].data)
 
             self.data['jvec'].loc[src, :, freq] = resp